third_party/WebKit/LayoutTests/batterystatus/add-listener-from-callback-expected.txt

CONSOLE MESSAGE: line 42: Uncaught TypeError: Cannot call method 'addEventListener' of undefined
Tests that adding a new event listener from a callback works as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
